Add separat cmakelists.

+++ src/joint_state.cpp
@@ -6,7 +6,7 @@
 // Set the joint angles, ensuring they are within the joint limits
 void JointState::setJointAngles(const std::vector<double>& angles) {
   // Check if the angles are within the joint limits
-  for (int i = 0; i < angles.size(); ++i) {
+  for (size_t i = 0; i < angles.size(); ++i) {
     const Joint& joint = robot_config_.getJoints().at(i);
     if (angles.at(i) < joint.theta_min || angles.at(i) > joint.theta_max) {
       std::ostringstream msg;
